requirements.txt
progress
ipython
graphs
dm-sonnet<2
sklearn
scipy
numpy
Pillow
tensorflow==1.15
tqdm
